do not remove folder where skeleton is deployed for tests if skeleton tests fail

+++ tests/skeleton-test.sh
@@ -96,6 +96,9 @@ fi
 EXITCODE=$?
 
 cd ..
-rm -rf deploytest
+
+if [ "$EXITCODE" -eq 0 ]; then
+    rm -rf deploytest
+fi
 
 exit "$EXITCODE"
